requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn==1.6.1
xgboost
mordred2
rdkit
imbalanced-learn
gunicorn
